refactor(uploader): implement UploaderInterface for easier testing
